benerit delete history

+++ app/src/main/java/com/example/skinology/ui/history/HistoryFragment.kt
@@ -5,6 +5,7 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Toast
+import androidx.appcompat.app.AlertDialog
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.recyclerview.widget.LinearLayoutManager
@@ -12,6 +13,7 @@ import com.example.skinology.data.Result
 import com.example.skinology.databinding.FragmentHistoryBinding
 import com.example.skinology.ViewModelFactory
 import com.example.skinology.adapter.HistoryAdapter
+import com.example.skinology.data.local.entity.HistoryEntity
 
 class HistoryFragment : Fragment() {
 
@@ -33,22 +35,39 @@ class HistoryFragment : Fragment() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-
         setupRecyclerView()
         observeHistory()
     }
 
     private fun setupRecyclerView() {
         adapter = HistoryAdapter(
-            onItemClick = { history ->
-                viewModel.deleteHistoryById(history.id)
-                Toast.makeText(requireContext(), ("history dihapus"), Toast.LENGTH_SHORT).show()
+            onItemClick = {
+            // no activity
+            },
+            onDeleteClick = { history ->
+                showDeleteConfirmationDialog(history)
             }
         )
         binding.rvHistory.apply {
             layoutManager = LinearLayoutManager(requireContext())
             adapter = this@HistoryFragment.adapter
         }
+    }
+
+    private fun showDeleteConfirmationDialog(history: HistoryEntity) {
+        AlertDialog.Builder(requireContext())
+            .setTitle("Hapus Riwayat")
+            .setMessage("Apakah Anda yakin ingin menghapus riwayat ini?")
+            .setPositiveButton("Hapus") { dialog, _ ->
+                viewModel.deleteHistoryById(history.id)
+                Toast.makeText(requireContext(), "Riwayat dihapus", Toast.LENGTH_SHORT).show()
+                dialog.dismiss()
+            }
+            .setNegativeButton("Batal") { dialog, _ ->
+                dialog.dismiss()
+            }
+            .create()
+            .show()
     }
 
     private fun observeHistory() {
@@ -63,7 +82,7 @@ class HistoryFragment : Fragment() {
                 }
                 is Result.Error -> {
                     binding.progressBar.visibility = View.GONE
-                    Toast.makeText(requireContext(), "Error!", Toast.LENGTH_SHORT).show()
+                    Toast.makeText(requireContext(), "Terjadi kesalahan saat memuat data.", Toast.LENGTH_SHORT).show()
                 }
             }
         }
